Typprüfung für die Helper Methoden
Foundation nutzt NSUInteger Typedefs

+++ iBroker/Common/Helper.h
@@ -10,7 +10,7 @@
 #import <Cocoa/Cocoa.h>
 
 @interface Helper : NSObject
-+ (NSString*) double2German:(double) value min:(int) min max:(int) max;
-+ (NSString*) double2GermanPercent:(double) value fractions:(int) fractions;
++ (NSString*) double2German:(double) value min:(NSUInteger) min max:(NSUInteger) max;
++ (NSString*) double2GermanPercent:(double) value fractions:(NSUInteger) fractions;
 + (NSModalResponse)messageText:(NSString*) message info:(NSString*) info;
 @end
